+++ src/test/kotlin/com/osscube/api/domain/service/open_source/GetOpenSourcesTest.kt
@@ -1,12 +1,10 @@
-package com.osscube.api.domain.service
+package com.osscube.api.domain.service.open_source
 
 import com.osscube.api.config.TestContainers
-import com.osscube.api.domain.dto.OpenSourceSaveRequestDto
-import com.osscube.api.domain.exception.open_source.OpenSourceAlreadyExistsException
 import com.osscube.api.domain.model.entity.OpenSource
 import com.osscube.api.domain.model.repository.OpenSourceRepository
+import com.osscube.api.domain.service.OpenSourceService
 import org.assertj.core.api.Assertions.assertThat
-import org.assertj.core.api.Assertions.assertThatThrownBy
 import org.assertj.core.api.Assertions.tuple
 import org.junit.jupiter.api.AfterEach
 import org.junit.jupiter.api.DisplayName
@@ -15,7 +13,7 @@ import org.springframework.beans.factory.annotation.Autowired
 import org.springframework.boot.test.context.SpringBootTest
 
 @SpringBootTest
-class OpenSourceServiceTest : TestContainers() {
+class GetOpenSourcesTest : TestContainers() {
     @Autowired
     private lateinit var openSourceService: OpenSourceService
 
@@ -25,39 +23,6 @@ class OpenSourceServiceTest : TestContainers() {
     @AfterEach
     fun cleansing() {
         openSourceRepository.deleteAllInBatch()
-    }
-
-    @DisplayName("오픈소스명, 오픈소스 출처를 받아서 오픈소스를 저장한다.")
-    @Test
-    fun saveOpenSource() {
-        // given
-        val name = "name"
-        val originUrl = "originUrl"
-        val requestDto = OpenSourceSaveRequestDto(name, originUrl)
-
-        // when
-        val responseDto = openSourceService.saveOpenSource(requestDto)
-
-        // then
-        assertThat(responseDto.openSourceId)
-            .hasSize(36)
-        assertThat(responseDto)
-            .extracting("name", "originUrl")
-            .contains(name, originUrl)
-    }
-
-    @DisplayName("이미 저장된 오픈소스를 다시 저장하면 예외가 발생한다.")
-    @Test
-    fun saveOpenSourceThrowsExceptionIfOpenSourceExists() {
-        // given
-        val requestDto = OpenSourceSaveRequestDto("name", "originUrl")
-
-        val openSource = OpenSource.of(requestDto)
-        openSourceRepository.save(openSource)
-
-        // when // then
-        assertThatThrownBy { openSourceService.saveOpenSource(requestDto) }
-            .isInstanceOf(OpenSourceAlreadyExistsException::class.java)
     }
 
     @DisplayName("모든 오픈소스를 조회한다.")
